Rename filter stats api by filter stats backend
