+++ lib/cartodb/api/api-router.js
@@ -16,7 +16,7 @@ const AuthBackend = require('../backends/auth');
 
 const UserLimitsApi = require('../backends/api/user_limits_api');
 const OverviewsMetadataApi = require('../backends/api/overviews_metadata_api');
-const FilterStatsApi = require('../backends/api/filter_stats_api');
+const FilterStatsApi = require('../backends/filter-stats');
 const TablesExtentApi = require('../backends/api/tables_extent_api');
 
 const LayergroupAffectedTablesCache = require('../cache/layergroup_affected_tables');
@@ -124,7 +124,7 @@ module.exports = class ApiRouter {
         const pgQueryRunner = new PgQueryRunner(pgConnection);
         const overviewsMetadataApi = new OverviewsMetadataApi(pgQueryRunner);
 
-        const filterStatsApi = new FilterStatsApi(pgQueryRunner);
+        const filterStatsBackend = new FilterStatsApi(pgQueryRunner);
         const tablesExtentApi = new TablesExtentApi(pgQueryRunner);
 
         const mapConfigAdapter = new MapConfigAdapter(
@@ -134,7 +134,7 @@ module.exports = class ApiRouter {
             new DataviewsWidgetsAdapter(),
             new AnalysisMapConfigAdapter(analysisBackend),
             new AggregationMapConfigAdapter(pgConnection),
-            new MapConfigOverviewsAdapter(overviewsMetadataApi, filterStatsApi),
+            new MapConfigOverviewsAdapter(overviewsMetadataApi, filterStatsBackend),
             new TurboCartoAdapter()
         );
