+++ lib/cartodb/backends/filter-stats.js
@@ -1,12 +1,12 @@
 var _ = require('underscore');
 var step = require('step');
-var AnalysisFilter = require('../../models/filter/analysis');
+var AnalysisFilter = require('../models/filter/analysis');
 
-function FilterStatsApi(pgQueryRunner) {
+function FilterStatsBackends(pgQueryRunner) {
     this.pgQueryRunner = pgQueryRunner;
 }
 
-module.exports = FilterStatsApi;
+module.exports = FilterStatsBackends;
 
 function getEstimatedRows(pgQueryRunner, username, query, callback) {
     pgQueryRunner.run(username, "EXPLAIN (FORMAT JSON)"+query, function(err, result_rows) {
@@ -23,7 +23,7 @@ function getEstimatedRows(pgQueryRunner, username, query, callback) {
     });
 }
 
-FilterStatsApi.prototype.getFilterStats = function (username, unfiltered_query, filters, callback) {
+FilterStatsBackends.prototype.getFilterStats = function (username, unfiltered_query, filters, callback) {
   var stats = {};
   var self = this;
   step(
