+++ lib/cartodb/models/mapconfig/adapter/mapconfig-overviews-adapter.js
@@ -2,9 +2,9 @@ var step = require('step');
 var queue = require('queue-async');
 var _ = require('underscore');
 
-function MapConfigOverviewsAdapter(overviewsMetadataApi, filterStatsApi) {
+function MapConfigOverviewsAdapter(overviewsMetadataApi, filterStatsBackend) {
     this.overviewsMetadataApi = overviewsMetadataApi;
-    this.filterStatsApi = filterStatsApi;
+    this.filterStatsBackend = filterStatsBackend;
 }
 
 module.exports = MapConfigOverviewsAdapter;
@@ -53,7 +53,7 @@ MapConfigOverviewsAdapter.prototype.getMapConfig = function(user, requestMapConf
                    function collectStatsData(err, filters, unfiltered_query) {
                        var next_step = this;
                        if ( filters ) {
-                           self.filterStatsApi.getFilterStats(
+                           self.filterStatsBackend.getFilterStats(
                                user,
                                unfiltered_query, filters,
                                function(err, stats) {
